+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+- Add dual-clock generator.
+
+
 ## v0.2.0 - 2019-08-20
 
 ### Added